+++ services/films_service.py
@@ -5,16 +5,30 @@ class FilmService:
     films_repository = FilmsRepositorySQLAlchemy()
 
     def get_all(self):
-        return self.films_repository.get_all()
+        return {"films": [item.to_dict(
+                only=("id", "title", "release_date", "duration", "director", "genre_id"))
+                for item in self.films_repository.get_all()]}
 
     def get_film(self, film_id):
-        return [self.films_repository.get_film(film_id)]
+        film = self.films_repository.get_film(film_id)
+        if film:
+            return {"film": [item.to_dict(
+                    only=("id", "title", "release_date", "duration", "director", "genre_id"))
+                    for item in [film]]}
+        return {"error": "not found"}
 
     def add(self, film):
-        return self.films_repository.add(film)
+        self.films_repository.add(film)
+        return {"success": "ok"}
 
     def update(self, film_id, film):
-        return self.films_repository.update(film_id, film)
+        if self.films_repository.get_film(film_id):
+            self.films_repository.update(film_id, film)
+            return {"success": "ok"}
+        return {"error": "not found"}
 
     def delete(self, film_id):
-        return self.films_repository.delete(film_id)
+        if self.films_repository.get_film(film_id):
+            self.films_repository.delete(film_id)
+            return {"success": "ok"}
+        return {"error": "not found"}
