+++ services/halls_service.py
@@ -8,13 +8,24 @@ class HallService:
         return self.halls_repository.get_all()
 
     def get_hall(self, hall_id):
-        return [self.halls_repository.get_hall(hall_id)]
+        hall = self.halls_repository.get_hall(hall_id)
+        if hall:
+            return {"hall": [item.to_dict(only=("id", "title", "cinema_id"))
+                             for item in [hall]]}
+        return {"success": "ok"}
 
     def add(self, hall):
-        return self.halls_repository.add(hall)
+        self.halls_repository.add(hall)
+        return {"success": "ok"}
 
     def update(self, hall_id, hall):
-        return self.halls_repository.update(hall_id, hall)
+        if self.halls_repository.get_hall(hall_id):
+            self.halls_repository.update(hall_id, hall)
+            return {"success": "ok"}
+        return {"error": "not found"}
 
     def delete(self, hall_id):
-        return self.halls_repository.delete(hall_id)
+        if self.halls_repository.get_hall(hall_id):
+            self.halls_repository.delete(hall_id)
+            return {"success": "ok"}
+        return {"error": "not found"}
