+++ services/user_service.py
@@ -5,16 +5,28 @@ class UserService:
     users_repository = UsersRepositorySQLAlchemy()
 
     def get_all(self):
-        return self.users_repository.get_all()
+        return {"users": [item.to_dict(only=("id", "login", "password", "email", "role_id"))
+                          for item in self.users_repository.get_all()]}
 
-    def get_user(self, user_id: int):
-        return [self.users_repository.get_user(user_id)]
+    def get_user(self, user_id):
+        user = self.users_repository.get_user(user_id)
+        if user:
+            return {"user": [item.to_dict(only=("id", "login", "password", "email", "role_id"))
+                             for item in [user]]}
+        return {"error": "not found"}
 
     def add(self, user):
-        return self.users_repository.add(user)
+        self.users_repository.add(user)
+        return {"success": "ok"}
 
     def update(self, user_id, new_user):
-        return self.users_repository.update(user_id, new_user)
+        if self.users_repository.get_user(user_id):
+            self.users_repository.update(user_id, new_user)
+            return {"success": "ok"}
+        return {"error": "not found"}
 
     def delete(self, user_id):
-        return self.users_repository.delete(user_id)
+        if self.users_repository.get_user(user_id):
+            self.users_repository.delete(user_id)
+            return {"success": "ok"}
+        return {"error": "not found"}
